Selection of Contrastive and Penalization Parameters

A wrapper function for fitting various internal functions to select the optimal setting of the contrastive and penalization parameters. For internal use only.

selectParams(target, background, center, scale, n_eigen, contrasts,
  penalties, clust_method, n_centers, max_iter, n_medoids, parallel)

Arguments

target
The target (experimental) data set, in a standard format such as a data.frame or matrix.

background
The background data set, in a standard format such as a data.frame or matrix. Note that the number of features must match the number of features in the target data.

center
A logical indicating whether the target and background data sets should be centered to mean zero.

scale
A logical indicating whether the target and background data sets should be scaled to unit variance.

n_eigen
A numeric indicating the number of eigenvectors (or sparse contrastive components) to be computed. The default is to compute two such eigenvectors.

contrasts
A numeric vector of the contrastive parameters. Each element must be a unique non-negative real number. The default is to use 40 logarithmically spaced values between 0.1 and 1000.

penalties
A numeric vector of the L1 penalty terms on the loadings. The default is to use 20 equidistant values between 0.05 and 1.

clust_method
A character specifying the clustering method to use for choosing the optimal constrastive parameter. Currently, this is limited to either k-means or partitioning around medoids (PAM). The default is k-means clustering.

n_centers
A numeric giving the number of centers to use in the clustering algorithm. If set to 1, cPCA, as first proposed by Abid et al., is performed, regardless of what the penalties argument is set to.

max_iter
A numeric giving the maximum number of iterations to be used in k-means clustering, defaulting to 10.

n_medoids
A numeric indicating the number of medoids to consider if n_centers is set to 1. The default is 8 such medoids.

parallel
A logical indicating whether to invoke parallel processing via the BiocParallel infrastructure. The default is FALSE for sequential evaluation.

Value

Output structure matching either that of fitCPCA or fitGrid (or their parallelized variants, namely either bpFitCPCA and link{bpFitGrid}, respectively).